rename module and package name `keyvaults` -> `keyvault
