+++ azure-toolkit-libs/azure-toolkit-keyvault-lib/src/main/java/com/microsoft/azure/toolkit/lib/keyvault/AzureKeyVault.java
@@ -1,0 +1,79 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+
+package com.microsoft.azure.toolkit.lib.keyvault;
+
+import com.azure.core.http.policy.HttpLogDetailLevel;
+import com.azure.core.http.policy.HttpLogOptions;
+import com.azure.core.http.policy.UserAgentPolicy;
+import com.azure.core.management.profile.AzureProfile;
+import com.azure.resourcemanager.keyvault.KeyVaultManager;
+import com.azure.resourcemanager.resources.ResourceManager;
+import com.azure.resourcemanager.resources.fluentcore.policy.ProviderRegistrationPolicy;
+import com.azure.resourcemanager.resources.models.Providers;
+import com.microsoft.azure.toolkit.lib.Azure;
+import com.microsoft.azure.toolkit.lib.AzureConfiguration;
+import com.microsoft.azure.toolkit.lib.auth.Account;
+import com.microsoft.azure.toolkit.lib.auth.AzureAccount;
+import com.microsoft.azure.toolkit.lib.common.model.AbstractAzService;
+import com.microsoft.azure.toolkit.lib.common.model.AbstractAzServiceSubscription;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.util.Optional;
+
+public class AzureKeyVault extends AbstractAzService<KeyVaultSubscription, KeyVaultManager> {
+    public AzureKeyVault() {
+        super("Microsoft.KeyVault");
+    }
+
+    @Nonnull
+    public KeyVaultModule keyVaults(@Nonnull String subscriptionId) {
+        final KeyVaultSubscription rm = get(subscriptionId, null);
+        assert rm != null;
+        return rm.getKeyVaultModule();
+    }
+
+    @Nonnull
+    @Override
+    protected KeyVaultSubscription newResource(@Nonnull KeyVaultManager keyVaultManager) {
+        return new KeyVaultSubscription(keyVaultManager.serviceClient().getSubscriptionId(), this);
+    }
+
+    @Nullable
+    @Override
+    protected KeyVaultManager loadResourceFromAzure(@Nonnull String subscriptionId, String resourceGroup) {
+        final Account account = Azure.az(AzureAccount.class).account();
+        final String tenantId = account.getSubscription(subscriptionId).getTenantId();
+        final AzureConfiguration config = Azure.az().config();
+        final String userAgent = config.getUserAgent();
+        final HttpLogOptions logOptions = new HttpLogOptions();
+        logOptions.setLogLevel(Optional.ofNullable(config.getLogLevel()).map(HttpLogDetailLevel::valueOf).orElse(HttpLogDetailLevel.NONE));
+        final AzureProfile azureProfile = new AzureProfile(tenantId, subscriptionId, account.getEnvironment());
+        final Providers providers = ResourceManager.configure()
+            .withHttpClient(AbstractAzServiceSubscription.getDefaultHttpClient())
+            .withPolicy(new UserAgentPolicy(userAgent))
+            .authenticate(account.getTokenCredential(subscriptionId), azureProfile)
+            .withSubscription(subscriptionId).providers();
+        return KeyVaultManager
+            .configure()
+            .withHttpClient(AbstractAzServiceSubscription.getDefaultHttpClient())
+            .withLogOptions(logOptions)
+            .withPolicy(new UserAgentPolicy(userAgent))
+            .withPolicy(new ProviderRegistrationPolicy(providers)) // add policy to auto register resource providers
+            .authenticate(account.getTokenCredential(subscriptionId), azureProfile);
+    }
+
+    @Nonnull
+    @Override
+    public String getResourceTypeName() {
+        return "Key Vaults";
+    }
+
+    @Nonnull
+    public String getServiceNameForTelemetry() {
+        return "keyvaults";
+    }
+}
