--- azure-toolkit-libs/azure-toolkit-keyvaults-lib/src/main/resources/META-INF/services/com.microsoft.azure.toolkit.lib.AzService
@@ -1,1 +1,0 @@
-com.microsoft.azure.toolkit.lib.keyvaults.AzureKeyVault
